backend/requirements.txt
fastapi
uvicorn
python-pptx
matplotlib
pillow
python-multipart
aiofiles
requests
# For copywriter_agent (PEFT + transformers must be compatible)
torch
transformers>=4.45.0
peft
accelerate
google-generativeai
duckduckgo-search
ddgs